+++ frontend/src/lib/components/accounts/SelectAccountDropdown.svelte
@@ -1,14 +1,13 @@
 <script lang="ts">
-  import { nnsAccountsListStore } from "$lib/derived/accounts-list.derived";
   import { i18n } from "$lib/stores/i18n";
   import type { Account } from "$lib/types/account";
   import {
     getAccountByRootCanister,
     getAccountsByRootCanister,
   } from "$lib/utils/accounts.utils";
-  import { Dropdown, DropdownItem } from "@dfinity/gix-components";
-  import { snsAccountsStore } from "$lib/stores/sns-accounts.store";
+  import { Dropdown, DropdownItem, Spinner } from "@dfinity/gix-components";
   import type { Principal } from "@dfinity/principal";
+  import { universesAccountsStore } from "$lib/derived/universes-accounts.derived";
 
   export let selectedAccount: Account | undefined = undefined;
   export let rootCanisterId: Principal;
@@ -18,32 +17,25 @@
   // To avoid cyclical dependencies, we don't update this if `selectedAccount` changes
   let selectedAccountIdentifier: string | undefined =
     selectedAccount?.identifier;
+
   $: selectedAccount = getAccountByRootCanister({
     identifier: selectedAccountIdentifier,
     rootCanisterId,
-    nnsAccounts: $nnsAccountsListStore,
-    snsAccounts: $snsAccountsStore,
+    universesAccounts: $universesAccountsStore,
   });
 
+  let selectableAccounts: Account[] = [];
   $: selectableAccounts =
     getAccountsByRootCanister({
       rootCanisterId,
-      nnsAccounts: $nnsAccountsListStore,
-      snsAccounts: $snsAccountsStore,
+      universesAccounts: $universesAccountsStore,
     })?.filter(filterAccounts) ?? [];
 </script>
 
 {#if selectableAccounts.length === 0}
-  <Dropdown
-    name="account"
-    disabled
-    selectedValue="no-accounts"
-    testId="select-account-dropdown"
-  >
-    <DropdownItem value="no-accounts">
-      {$i18n.accounts.no_account_select}
-    </DropdownItem>
-  </Dropdown>
+  <div class="select">
+    <Spinner size="small" inline />
+  </div>
 {:else}
   <Dropdown
     name="account"
@@ -57,3 +49,18 @@
     {/each}
   </Dropdown>
 {/if}
+
+<style lang="scss">
+  @use "@dfinity/gix-components/styles/mixins/form";
+  .select {
+    @include form.input;
+
+    position: relative;
+    box-sizing: border-box;
+
+    padding: var(--padding-2x);
+    border-radius: var(--border-radius);
+
+    width: var(--dropdown-width, auto);
+  }
+</style>