+++ frontend/src/lib/components/common/MaxButton.svelte
@@ -3,7 +3,12 @@
   import { i18n } from "$lib/stores/i18n";
 </script>
 
-<button on:click|preventDefault class="ghost" type="button">
+<button
+  data-tid="max-button"
+  on:click|preventDefault
+  class="ghost"
+  type="button"
+>
   <span class="icon">
     <IconSubdirectory />
   </span>