+++ frontend/src/lib/components/neuron-detail/NeuronJoinFundCard.svelte
@@ -22,11 +22,9 @@
     <KeyValuePairInfo testId="join-community-fund">
       <h3 slot="key">{$i18n.neurons.community_fund_title}</h3>
 
-      <svelte:fragment slot="info"
-        ><div class="info">
-          <Html text={$i18n.neuron_detail.community_fund_more_info} />
-        </div></svelte:fragment
-      >
+      <div class="info" slot="info">
+        <Html text={$i18n.neuron_detail.community_fund_more_info} />
+      </div>
     </KeyValuePairInfo>
 
     <div class="join">
@@ -42,18 +40,12 @@
     line-height: var(--line-height-standard);
   }
 
+  .info {
+    line-height: initial;
+  }
+
   .join {
     --checkbox-label-order: 1;
     --checkbox-padding: var(--padding) 0;
-  }
-
-  .info {
-    // For the link inside "i18n.neuron_detail.community_fund_more_info"
-    :global(a) {
-      color: var(--primary);
-      text-decoration: none;
-      font-size: inherit;
-      line-height: inherit;
-    }
   }
 </style>