+++ frontend/src/lib/components/project-detail/ProjectTimeline.svelte
@@ -6,9 +6,13 @@
   } from "$lib/types/project-detail.context";
   import type { SnsSummary, SnsSummarySwap } from "$lib/types/sns";
   import { i18n } from "$lib/stores/i18n";
-  import { durationTillSwapDeadline } from "$lib/utils/projects.utils";
+  import {
+    durationTillSwapDeadline,
+    durationTillSwapStart,
+  } from "$lib/utils/projects.utils";
   import { secondsToDuration } from "$lib/utils/date.utils";
   import { Value, KeyValuePair } from "@dfinity/gix-components";
+  import { SnsSwapLifecycle } from "@dfinity/sns";
 
   const { store: projectDetailStore } = getContext<ProjectDetailContext>(
     PROJECT_DETAIL_CONTEXT_KEY
@@ -20,9 +24,12 @@
 
   let durationTillDeadline: bigint | undefined;
   $: durationTillDeadline = durationTillSwapDeadline(swap);
+
+  let durationTillStart: bigint | undefined;
+  $: durationTillStart = durationTillSwapStart(swap);
 </script>
 
-{#if durationTillDeadline !== undefined}
+{#if durationTillDeadline && swap.lifecycle === SnsSwapLifecycle.Open}
   <KeyValuePair>
     <span slot="key">
       {$i18n.sns_project_detail.deadline}
@@ -32,4 +39,14 @@
     </Value>
   </KeyValuePair>
 {/if}
-<div />
+
+{#if durationTillStart && swap.lifecycle === SnsSwapLifecycle.Adopted}
+  <KeyValuePair>
+    <span slot="key">
+      {$i18n.sns_project_detail.starts}
+    </span>
+    <Value slot="value">
+      {secondsToDuration(durationTillStart)}
+    </Value>
+  </KeyValuePair>
+{/if}