+++ frontend/src/lib/pages/Launchpad.svelte
@@ -3,15 +3,39 @@
   import Proposals from "$lib/components/launchpad/Proposals.svelte";
   import { i18n } from "$lib/stores/i18n";
   import { SnsSwapLifecycle } from "@dfinity/sns";
-  import { committedProjectsStore } from "$lib/derived/projects.derived";
+  import {
+    snsProjectsAdoptedStore,
+    snsProjectsCommittedStore,
+  } from "$lib/derived/sns/sns-projects.derived";
+  import { loadSnsSwapCommitments } from "$lib/services/sns.services";
+  import { authStore } from "$lib/stores/auth.store";
+  import { isSignedIn } from "$lib/utils/auth.utils";
+
+  const loadSnsSale = async () => {
+    if (!isSignedIn($authStore.identity)) {
+      return;
+    }
+
+    await loadSnsSwapCommitments();
+  };
+
+  $: $authStore.identity, (async () => await loadSnsSale())();
 
   let showCommitted = false;
-  $: showCommitted = ($committedProjectsStore?.length ?? []) > 0;
+  $: showCommitted = ($snsProjectsCommittedStore?.length ?? []) > 0;
+
+  let showAdopted = false;
+  $: showAdopted = ($snsProjectsAdoptedStore?.length ?? []) > 0;
 </script>
 
 <main>
   <h2>{$i18n.sns_launchpad.open_projects}</h2>
   <Projects status={SnsSwapLifecycle.Open} />
+
+  {#if showAdopted}
+    <h2>{$i18n.sns_launchpad.upcoming_projects}</h2>
+    <Projects status={SnsSwapLifecycle.Adopted} />
+  {/if}
 
   {#if showCommitted}
     <h2>{$i18n.sns_launchpad.committed_projects}</h2>