+++ frontend/src/lib/services/ckbtc-minter.services.ts
@@ -1,0 +1,7 @@
+import { getBTCAddress as getBTCAddressAPI } from "$lib/api/ckbtc-minter.api";
+import { getAuthenticatedIdentity } from "$lib/services/auth.services";
+
+export const getBTCAddress = async (): Promise<string> => {
+  const identity = await getAuthenticatedIdentity();
+  return getBTCAddressAPI({ identity });
+};