+++ frontend/src/routes/(app)/(u)/(detail)/wallet/+layout.svelte
@@ -3,12 +3,17 @@
   import Content from "$lib/components/layout/Content.svelte";
   import { goto } from "$app/navigation";
   import { accountsPathStore } from "$lib/derived/paths.derived";
+  import { isNullish } from "$lib/utils/utils";
+  import { navigating } from "$app/stores";
 
   const back = (): Promise<void> => goto($accountsPathStore);
 </script>
 
-<Layout>
-  <Content {back}>
-    <slot />
-  </Content>
-</Layout>
+<!-- Workaround for SvelteKit issue https://github.com/sveltejs/kit/issues/5434 -->
+{#if isNullish($navigating)}
+  <Layout>
+    <Content {back}>
+      <slot />
+    </Content>
+  </Layout>
+{/if}