+++ frontend/src/tests/lib/pages/ProjectDetail.spec.ts
@@ -25,12 +25,6 @@ jest.mock("$lib/services/sns.services", () => {
   };
 });
 
-jest.mock("$lib/services/sns.services", () => {
-  return {
-    loadSnsSwapCommitments: jest.fn().mockResolvedValue(Promise.resolve()),
-  };
-});
-
 describe("ProjectDetail", () => {
   describe("present project in store", () => {
     page.mock({ data: { universe: null } });
@@ -60,16 +54,16 @@ describe("ProjectDetail", () => {
       rootCanisterId: mockSnsFullProject.rootCanisterId.toText(),
     };
 
-    it("should load summary", () => {
+    it("should load summary", async () => {
       render(ProjectDetail, props);
 
-      waitFor(() => expect(loadSnsSummary).toBeCalled());
+      await waitFor(() => expect(loadSnsSummary).toBeCalled());
     });
 
-    it("should load swap state", () => {
+    it("should load swap state", async () => {
       render(ProjectDetail, props);
 
-      waitFor(() => expect(loadSnsSwapCommitment).toBeCalled());
+      await waitFor(() => expect(loadSnsSwapCommitment).toBeCalled());
     });
 
     it("should render info section", async () => {