+++ frontend/src/tests/lib/services/sns-accounts-balance.services.spec.ts
@@ -1,12 +1,16 @@
 import * as ledgerApi from "$lib/api/sns-ledger.api";
-import { projectsAccountsBalance } from "$lib/derived/projects-accounts-balance.derived";
+import { universesAccountsBalance } from "$lib/derived/universes-accounts-balance.derived";
 import * as services from "$lib/services/sns-accounts-balance.services";
 import { snsAccountsStore } from "$lib/stores/sns-accounts.store";
 import { toastsError } from "$lib/stores/toasts.store";
+import { tokensStore } from "$lib/stores/tokens.store";
 import { tick } from "svelte";
 import { get } from "svelte/store";
 import { mockSnsMainAccount } from "../../mocks/sns-accounts.mock";
-import { mockSnsSummaryList } from "../../mocks/sns-projects.mock";
+import {
+  mockSnsSummaryList,
+  mockSnsToken,
+} from "../../mocks/sns-projects.mock";
 
 jest.mock("$lib/stores/toasts.store", () => {
   return {
@@ -19,6 +23,7 @@ describe("sns-accounts-balance.services", () => {
     jest.clearAllMocks();
 
     snsAccountsStore.reset();
+    tokensStore.reset();
   });
 
   const summary = {
@@ -28,6 +33,10 @@ describe("sns-accounts-balance.services", () => {
   };
 
   it("should call api.getSnsAccounts and load balance in store", async () => {
+    jest
+      .spyOn(ledgerApi, "getSnsToken")
+      .mockImplementation(() => Promise.resolve(mockSnsToken));
+
     const spyQuery = jest
       .spyOn(ledgerApi, "getSnsAccounts")
       .mockImplementation(() => Promise.resolve([mockSnsMainAccount]));
@@ -38,12 +47,37 @@ describe("sns-accounts-balance.services", () => {
 
     await tick();
 
-    const store = get(projectsAccountsBalance);
-    // Nns + 1 Sns
-    expect(Object.keys(store)).toHaveLength(2);
+    const store = get(universesAccountsBalance);
+    // Nns + ckBTC + 1 Sns
+    expect(Object.keys(store)).toHaveLength(3);
     expect(store[summary.rootCanisterId.toText()].balance.toE8s()).toEqual(
       mockSnsMainAccount.balance.toE8s()
     );
+    expect(spyQuery).toBeCalled();
+  });
+
+  it("should call api.getSnsToken and load it in store", async () => {
+    const spyQuery = jest
+      .spyOn(ledgerApi, "getSnsToken")
+      .mockImplementation(() => Promise.resolve(mockSnsToken));
+
+    jest
+      .spyOn(ledgerApi, "getSnsAccounts")
+      .mockImplementation(() => Promise.resolve([mockSnsMainAccount]));
+
+    await services.uncertifiedLoadSnsAccountsBalances({
+      rootCanisterIds: [mockSnsMainAccount.principal],
+    });
+
+    await tick();
+
+    const store = get(tokensStore);
+
+    expect(store[mockSnsMainAccount.principal.toText()]).toEqual({
+      token: mockSnsToken,
+      certified: false,
+    });
+
     expect(spyQuery).toBeCalled();
   });